update build some
cmake-modules-webos=19
move order of build files around, so webkit builds as early in process
as we can, so the layer will remain hopefully unchanged as much as possible
so that Docker doesn't rebuild it unnecessarily while working on other
pierces.
Figured out how to manually get luna-universalsearchmgr to build, and
get services setup.  Need to patch luna-usm around line 225 .. not sure
what is going wrong with the services, suspect it may be environment related
as it works from shell, but not from Docker creation.